feat: rename files from filter to sorting
